Add code refactoring to reduce code duplication
